+++ gkraken/conf.py
@@ -19,7 +19,7 @@ from typing import Dict, Any
 APP_PACKAGE_NAME = "gkraken"
 APP_NAME = "GKraken"
 APP_ID = "com.leinardi.gkraken"
-APP_VERSION = "0.8.0"
+APP_VERSION = "0.9.0"
 APP_ICON_NAME = APP_PACKAGE_NAME + ".png"
 APP_DB_NAME = APP_PACKAGE_NAME + ".db"
 APP_UI_NAME = APP_PACKAGE_NAME + ".glade"